# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OLAP Benchmark API — run benchmarks, stream logs, retrieve results
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: benchmarks
    description: Benchmark execution and results
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /benchmarks/status:
    get:
      operationId: getBenchmarkStatus
      tags: [benchmarks]
      summary: Get benchmark run status
      description: Returns which benchmarks have completed results and whether one is currently running
      responses:
        "200":
          description: Current status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenchmarkStatus"

  /benchmarks/results/{useCase}:
    get:
      operationId: getBenchmarkResults
      tags: [benchmarks]
      summary: Get results for a use case
      description: Returns the stored JSON results for the given use case (if available)
      parameters:
        - name: useCase
          in: path
          required: true
          schema:
            type: string
            enum:
              - dashboards
              - complex_joins
              - variant_test
              - clustering
              - acid_integrity
      responses:
        "200":
          description: Benchmark results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenchmarkResults"
        "404":
          description: No results for this use case yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /benchmarks/run/{useCase}:
    post:
      operationId: runBenchmark
      tags: [benchmarks]
      summary: Trigger a benchmark run
      description: Spawns the Python benchmark script as a background process. Connect to /api/benchmarks/logs/{useCase} (SSE) to stream live output.
      parameters:
        - name: useCase
          in: path
          required: true
          schema:
            type: string
            enum:
              - dashboards
              - complex_joins
              - variant_test
              - clustering
              - acid_integrity
      responses:
        "200":
          description: Benchmark started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunBenchmarkResponse"
        "409":
          description: A benchmark is already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    BenchmarkStatus:
      type: object
      properties:
        running:
          type: boolean
          description: Whether a benchmark is currently running
        runningUseCase:
          type: ["string", "null"]
          description: The use case currently running (null if idle)
        completedUseCases:
          type: array
          items:
            type: string
          description: Use cases that have result files available
        availableSystems:
          type: array
          items:
            type: string
          description: Database systems detected as available (postgres, duckdb, spark)
      required:
        - running
        - runningUseCase
        - completedUseCases
        - availableSystems

    BenchmarkResults:
      type: object
      description: Raw benchmark results JSON — structure varies by use case but always includes a top-level `validation` block per system
      additionalProperties: {}

    RunBenchmarkResponse:
      type: object
      properties:
        started:
          type: boolean
        useCase:
          type: string
        message:
          type: string
        pid:
          type: ["integer", "null"]
          description: Process ID of the spawned Python script
      required:
        - started
        - useCase
        - message

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error