# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OLAP Benchmark API — run benchmarks, stream logs, retrieve results
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns which benchmarks have completed results and whether one is currently running
 * @summary Get benchmark run status
 */
export const GetBenchmarkStatusResponse = zod.object({
  running: zod.boolean().describe("Whether a benchmark is currently running"),
  runningUseCase: zod
    .string()
    .nullable()
    .describe("The use case currently running (null if idle)"),
  completedUseCases: zod
    .array(zod.string())
    .describe("Use cases that have result files available"),
  availableSystems: zod
    .array(zod.string())
    .describe(
      "Database systems detected as available (postgres, duckdb, spark)",
    ),
});

/**
 * Returns the stored JSON results for the given use case (if available)
 * @summary Get results for a use case
 */
export const GetBenchmarkResultsParams = zod.object({
  useCase: zod.enum([
    "dashboards",
    "complex_joins",
    "variant_test",
    "clustering",
    "acid_integrity",
  ]),
});

export const GetBenchmarkResultsResponse = zod
  .record(zod.string(), zod.unknown())
  .describe(
    "Raw benchmark results JSON — structure varies by use case but always includes a top-level `validation` block per system",
  );

/**
 * Spawns the Python benchmark script as a background process. Connect to /api/benchmarks/logs/{useCase} (SSE) to stream live output.
 * @summary Trigger a benchmark run
 */
export const RunBenchmarkParams = zod.object({
  useCase: zod.enum([
    "dashboards",
    "complex_joins",
    "variant_test",
    "clustering",
    "acid_integrity",
  ]),
});

export const RunBenchmarkResponse = zod.object({
  started: zod.boolean(),
  useCase: zod.string(),
  message: zod.string(),
  pid: zod
    .number()
    .nullish()
    .describe("Process ID of the spawned Python script"),
});